@@ -1,3 +1,5 @@
 module.exports = {
-  tabWidth: 4
+    tabWidth: 4,
+    printWidth: 80,
+    proseWrap: "always",
 };
