@@ -1,5 +1,6 @@
 import { Component } from '@angular/core';
 import { LoaderService } from './providers/loader/loader.service';
+import { Team } from "@models";
 
 @Component({
   selector: 'app-root',
@@ -9,7 +10,7 @@ import { LoaderService } from './providers/loader/loader.service';
 export class AppComponent {
   title = 'Web';
 
-  constructor(public loader: LoaderService){
-    
+  constructor(public loader: LoaderService) {
+
   }
 }
